requirements.txt
semgrep>=1.0.0
jinja2>=3.0.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.26.0
typing-extensions>=4.0.0
pathlib>=1.0.1
gitpython>=3.1.0
# Gitleaks, Syft, Trivy must be installed separately as CLI tools (not pip)